[modify] refine log producer sample
